Download IMF data using Quandl API

A wrapper around Quandl. Downloads macroeconomic data for several indicators and countries covered by the IMF.

fq_imf(countries, indicators, verbose = TRUE, ...)

Arguments

countries
A vector or a list of character strings.

indicators
A vector or a list of character strings.

verbose
Should warning messages be printed? Default is TRUE.

...
Additional arguments to be passed into Quandl function.

Value

A tidy tibble with four columns: date, country, indicator and value.

Details

The countries argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or country names, but not both).

Sometimes the user may be interested in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

'ae' - Advanced Economics

'asean5' - Asean Top 5

'cis' - Commonwealth and Independent States

'eda' - Emerging and Developing Asia

'ede' - Emerging and Developing Economies

'edeuro' - Emerging and Developing Europe

'euro' - Euro Area

'eu' - European Union

'g7' - G7

'latam' - Latin America & Caribbean

'me' - Middle East

'oae' - Other Advanced Economies

'ssa' - Sub Saharan Africa

For any of those calls the fq_imf() will download data for all the countries in the requested region. A complete region list can be seen at: https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm.

The ... argument can be used to calibrate the query parameters. It accepts the following calls:

start_date: 'YYYY-MM-DD'

end_date: 'YYYY-MM-DD'

collapse: c("", "daily", "weekly", "monthly", "quarterly", "annual")

transform: c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")

The full list can be seen at: https://docs.quandl.com/docs/r.

Examples

# Download the US Output Gap
if (FALSE) {
  fq_imf('United States', 'NGAP_NPGDP')

# The example above is identical to
# fq_imf('USA', 'NGAP_NPGDP')

}