Download World Bank Data Using Quandl API

A wrapper around Quandl. Downloads macroeconomic data for several indicators and countries covered by World Bank.

fq_wb(countries, indicators, verbose = TRUE, ...)

Arguments

countries
A vector or a list of character strings.

indicators
A vector or a list of character strings.

verbose
Should warning messages be printed? Default is TRUE.

...
Additional arguments to be passed into Quandl function.

Value

A tidy tibble with four columns: date, country, indicator and value.

Details

The countries argument can be passed as an ISO code or as a country name. The only requirement is that the call must be consistent (must contain only ISO codes or country names, but not both).

Sometimes the user may be interesoted in downloading data for certain regions, like Europe, Latin America, Middle East, etc. For that reason, the countries argument also accepts the following calls:

'ae' - Advanced Economics

'asean5' - Asean Top 5

'cis' - Commonwealth and Independent States

'eda' - Emerging and Developing Asia

'ede' - Emerging and Developing Economies

'edeuro' - Emerging and Developing Europe

'euro' - Euro Area

'eu' - European Union

'g7' - G7

'latam' - Latin America & Caribbean

'me' - Middle East

'oae' - Other Advanced Economies

'ssa' - Sub Saharan Africa

For any of those calls the fq_wb() will download data for all the countries in the requested region. A complete region list can be seen at: https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/groups.htm.

The ... argument can be used to calibrate the query parameters. It accepts the following calls:

start_date: 'YYYY-MM-DD'

end_date: 'YYYY-MM-DD'

collapse: c("", "daily", "weekly", "monthly", "quarterly", "annual")

transform: c("", "diff", "rdiff", "normalize", "cumul", "rdiff_from")

The full list can be seen at: https://docs.quandl.com/docs/r.

Quandl has a diverse of World Bank datasets. The following items give a brief description of what you could expect to find when trying to dig further into each one of them:

World Development Indicators: Most current and accurate development indicators, compiled from officially-recognized international sources.

Worldwide Governance Indicators: Data on aggregate and individual governance indicators for six dimensions of governance.

Public Sector Debt: Data jointly developed by the World Bank and the International Monetary Fund, which brings together detailed public sector government debt data.

Poverty Statistics: Indicators on poverty headcount ratio, poverty gap, and number of poor at both international and national poverty lines.

Millennium Development Goals: Data drawn from the World Development Indicators, reorganized according to the goals and targets of the Millennium Development Goals (MDGs).

Jobs for Knowledge Platform: Indicators on labor-related topics.

International Development Association: Data on progress on aggregate outcomes for IDA (International Development Association) countries for selected indicators.

Health Nutrition and Population Statistics: Key health, nutrition and population statistics.

Global Findex (Global Financial Inclusion database): Indicators of financial inclusion measures on how people save, borrow, make payments and manage risk.

Global Financial Development: Data on financial system characteristics, including measures of size, use, access to, efficiency, and stability of financial institutions and markets.

GEP Economic Prospects: Data on the short-, medium, and long-term outlook for the global economy and the implications for developing countries and poverty reduction.

Gender Statistics: Data describing gender differences in earnings, types of jobs, sectors of work, farmer productivity, and entrepreneurs’ firm sizes and profits.

Global Economic Monitor: Data on global economic developments, with coverage of high-income, as well as developing countries.

Global Economic Monitor (GEM) Commodities: Data containing commodity prices and indices from 1960 to present.

Global Development Finance: Data on financial system characteristics, including measures of size, use, access to, efficiency, and stability of financial institutions and markets.

Enterprise Surveys: Company-level private sector data, covering business topics including finance, corruption, infrastructure, crime, competition, and performance measures.

Education Statistics: Internationally comparable indicators on education access, progression, completion, literacy, teachers, population, and expenditures.

Doing Business: Data on business regulations and their enforcement for member countries and selected cities at the subnational and regional level.

Corporate Scorecard: This database is designed to provide a strategic overview of the World Bank Group’s performance toward ending extreme poverty and promoting shared prosperity.

Africa Development Indicators: A collection of development indicators on Africa, including national, regional and global estimates.

Examples

if (FALSE) {
  # Download al indicators related to Rule of Law for the G7 countries from 2010 ownwards...
  fq_wb(countries  = 'g7',
        indicators = c('RL_EST', 'RL_NO_SRC'),
        start_date = '2010-01-01')
}