Search for the World Bank indicators avaiable in Quandl.

It covers 1309 indicators spread among the Governance, Development, Poverty and Public Sector databases.

fq_wb_search(...)

Arguments

...
Character strings to match the search.

Value

A tibble with the indicators' name and the respective Quandl code.

Details

The full list of indicators can be seen in the following links:

Governance: https://www.quandl.com/data/WWGI-World-Bank-Worldwide-Governance-Indicators/documentation/data-organization

Development: https://www.quandl.com/data/WWDI-World-Bank-World-Development-Indicators/documentation/data-organization

Poverty: https://www.quandl.com/data/WPOV-World-Bank-Poverty-Statistics/documentation/data-organization

Public Sector: https://www.quandl.com/data/WPSD-World-Bank-Public-Sector-Debt

Examples

# oil search
fq_wb_search(query = 'oil')

#> Error: Problem with `mutate()` input `download`.
#> x Timeout was reached: [s3.amazonaws.com] Connection timed out after 10002 milliseconds
#> i Input `download` is `purrr::map(...)`.

# The function is case insensitive
fq_wb_search(query = 'EDuCAtion')

#> Error: Problem with `mutate()` input `download`.
#> x Timeout was reached: [s3.amazonaws.com] Operation timed out after 10013 milliseconds with 0 out of 0 bytes received
#> i Input `download` is `purrr::map(...)`.