Search for the World Bank indicators avaiable in Quandl.

It covers 1309 indicators spread among the Governance, Development, Poverty and Public Sector databases.

fq_wb_search(...)

Arguments

...
Character strings to match the search.

Value

A tibble with the indicators' name and the respective Quandl code.

Details

The full list of indicators can be seen in the following links:

Governance: https://www.quandl.com/data/WWGI-World-Bank-Worldwide-Governance-Indicators/documentation/data-organization

Development: https://www.quandl.com/data/WWDI-World-Bank-World-Development-Indicators/documentation/data-organization

Poverty: https://www.quandl.com/data/WPOV-World-Bank-Poverty-Statistics/documentation/data-organization

Public Sector: https://www.quandl.com/data/WPSD-World-Bank-Public-Sector-Debt

Examples

# oil search
fq_wb_search(query = 'oil')

#> # A tibble: 45 x 3
#>    indicator                                           code            database 
#>    <chr>                                               <chr>           <chr>    
#>  1 CO2 intensity (kg per kg of oil equivalent energy … EN_ATM_CO2E_EG… developm…
#>  2 Combustible renewables and waste (metric tons of o… EG_USE_CRNW_KT… developm…
#>  3 Electricity production from oil sources (% of tota… EG_ELC_PETR_ZS  developm…
#>  4 Electricity production from oil sources (kWh)       EG_ELC_PETR_KH  developm…
#>  5 Electricity production from oil, gas and coal sour… EG_ELC_FOSL_ZS  developm…
#>  6 Energy production (kt of oil equivalent)            EG_EGY_PROD_KT… developm…
#>  7 Energy use (kg of oil equivalent per capita)        EG_USE_PCAP_KG… developm…
#>  8 Energy use (kg of oil equivalent) per $1,000 GDP (… EG_USE_COMM_GD… developm…
#>  9 Energy use (kt of oil equivalent)                   EG_USE_COMM_KT… developm…
#> 10 GDP per unit of energy use (constant 2005 PPP $ pe… EG_GDP_PUSE_KO… developm…
#> # … with 35 more rows

# The function is case insensitive
fq_wb_search(query = 'EDuCAtion')

#> # A tibble: 493 x 3
#>    indicator                                            code          database  
#>    <chr>                                                <chr>         <chr>     
#>  1 Adjusted savings: education expenditure (% of GNI)   NY_ADJ_AEDU_… developme…
#>  2 Adjusted savings: education expenditure (current US… NY_ADJ_AEDU_… developme…
#>  3 Labor force with primary education (% of total)      SL_TLF_PRIM_… developme…
#>  4 Labor force with primary education, female (% of fe… SL_TLF_PRIM_… developme…
#>  5 Labor force with primary education, male (% of male… SL_TLF_PRIM_… developme…
#>  6 Labor force with secondary education (% of total)    SL_TLF_SECO_… developme…
#>  7 Labor force with secondary education, female (% of … SL_TLF_SECO_… developme…
#>  8 Labor force with secondary education, male (% of ma… SL_TLF_SECO_… developme…
#>  9 Labor force with tertiary education (% of total)     SL_TLF_TERT_… developme…
#> 10 Labor force with tertiary education, female (% of f… SL_TLF_TERT_… developme…
#> # … with 483 more rows